+++ packages/thirdweb/src/contract/deployment/deploy-from-uri.ts
@@ -1,6 +1,6 @@
 import type { Prettify } from "../../utils/type-utils.js";
 import type { SharedDeployOptions } from "./types.js";
-import { fetchDeployMetadata } from "./utils/deploy-metadata.js";
+import { fetchDeployMetadata } from "../../utils/any-evm/deploy-metadata.js";
 import type { AbiConstructor } from "abitype";
 import { prepareDirectDeployTransaction } from "./deploy-with-abi.js";
 import { ensureBytecodePrefix } from "../../utils/bytecode/prefix.js";
