+++ packages/thirdweb/src/contract/verification/constructor-params.ts
@@ -1,7 +1,7 @@
 import type { Abi } from "abitype";
 import type { ThirdwebContract } from "../contract.js";
 import { fetchDeployBytecodeFromPublishedContractMetadata } from "./publisher.js";
-import { getCreate2FactoryAddress } from "../../utils/any-evm/create-2-factory.js";
+import { getCreate2FactoryAddress } from "../deployment/utils/create-2-factory.js";
 import { decodeAbiParameters } from "viem";
 import { getRpcClient } from "../../rpc/rpc.js";
 import { eth_getTransactionByHash } from "../../rpc/actions/eth_getTransactionByHash.js";
