+++ packages/thirdweb/src/utils/any-evm/compute-published-contract-deploy-info.ts
@@ -1,11 +1,11 @@
 import type { AbiConstructor } from "abitype";
-import type { ThirdwebClient } from "../../../client/client.js";
-import { getInitBytecodeWithSalt } from "../../../utils/any-evm/get-init-bytecode-with-salt.js";
+import type { ThirdwebClient } from "../../client/client.js";
+import { getInitBytecodeWithSalt } from "./get-init-bytecode-with-salt.js";
 import { fetchDeployMetadata } from "./deploy-metadata.js";
 import { fetchPublishedContract } from "./fetch-published-contract.js";
-import type { Chain } from "../../../chains/types.js";
-import { encodeAbiParameters } from "../../../utils/abi/encodeAbiParameters.js";
-import { getCreate2FactoryAddress } from "../../../utils/any-evm/create-2-factory.js";
+import type { Chain } from "../../chains/types.js";
+import { encodeAbiParameters } from "../abi/encodeAbiParameters.js";
+import { getCreate2FactoryAddress } from "../../contract/deployment/utils/create-2-factory.js";
 
 /**
  * @internal
